Remove code object files.  Copy contents to top of iot_pkcs11_mbedtls_c
Avoid disruptive project/cmake changes.
Uncrustify & move doxygen comments into PAL header.
